MOBILE-2993: Fix NSInvalidArgumentException in - [ProcessTasksCell updateStatusLabelUsingTask:]
Pooled tasks were being rendered in the "Tasks I Started" screen, they don't have an assignee which was causing the crash. Furthermore, a "Reassign" button was also available for this view but self.task is nil as the view is initialised with a process object. This change there also provides a better fix for MOBILE-2990.

Pooled tasks/workflows (Activiti and JBPM) are no longer displayed as there are other rendering issues (missing strings).